% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BVIAS_feed_pseudoherd.R
\name{f_BVIAS_feed_pseudoherd}
\alias{f_BVIAS_feed_pseudoherd}
\title{Function to estimate economic allocation ratios between herd outputs
\code{f_pseudoherd_feed_impact} Estimate herd outputs}
\usage{
f_BVIAS_feed_pseudoherd(object)
}
\arguments{
\item{object}{a FADN2Footprint object}
}
\value{
A list with, for each livestock category:
\itemize{
\item Q_obs: the number of animals observed in the farm (in number of heads)
\item Q_obs_pseudo: the number of animals of the pseudo-herd (in number of heads)
}
}
\description{
Function to estimate economic allocation ratios between herd outputs
\code{f_pseudoherd_feed_impact} Estimate herd outputs
}
\examples{
data(fadn_fict)
fadn_fict_obj = data_4FADN2Footprint(fadn_fict)
f_pseudoherd_feed_impact(object = fadn_fict_obj)

}
\concept{footprint-biodiv}
