% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice_crop_diversity.R
\name{f_crop_diversity}
\alias{f_crop_diversity}
\title{Function to estimate crop diversity based on FADN data
\code{crop_diversity} Create a tibble with the following variables:  'farm_id', 'land_use_type', 'Shannon', 'crop_nb_LU'}
\usage{
f_crop_diversity(object, overwrite = FALSE)
}
\arguments{
\item{object}{a FADN2Footprint object}

\item{overwrite}{Logical (default FALSE). If FALSE and cached results exist,
the function returns
the cached object and no recomputation is performed. If TRUE, existing
cached GHGE results are ignored and computations are re-run.}
}
\value{
A tibble with:
\itemize{
\item crop_nb_pLU: number of crop per land use type (i.e., arable or grassland)
\item Shannon: the Shannon index per land use type
\item R_Simpson: the reciprocal Simpson index per land use type
}
}
\description{
Function to estimate crop diversity based on FADN data
\code{crop_diversity} Create a tibble with the following variables:  'farm_id', 'land_use_type', 'Shannon', 'crop_nb_LU'
}
\examples{
data(fadn_fict)
fadn_fict_obj = data_4FADN2Footprint(fadn_fict)
crop_diversity(object = fadn_fict_obj)

}
\concept{practice-crop}
