% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice_ground_cover.R
\name{f_ground_cover}
\alias{f_ground_cover}
\title{Function to estimate ground cover based on FADN, LPIS and SENTINEL-2 data
\code{ground_cover} Create a tibble with the following variables:  'farm_id', 'land_use_type', 'ground_cover'}
\usage{
f_ground_cover(object, overwrite = FALSE)
}
\arguments{
\item{object}{a FADN2Footprint object}

\item{overwrite}{Logical (default FALSE). If FALSE and cached results exist,
the function returns
the cached object and no recomputation is performed. If TRUE, existing
cached GHGE results are ignored and computations are re-run.}
}
\value{
A tibble with:
\itemize{
\item ground_cover: average of hedge linear meter per hectare
}
}
\description{
Function to estimate ground cover based on FADN, LPIS and SENTINEL-2 data
\code{ground_cover} Create a tibble with the following variables:  'farm_id', 'land_use_type', 'ground_cover'
}
\examples{
data(fadn_fict)
fadn_fict_obj = data_4FADN2Footprint(fadn_fict)
ground_cover(object = fadn_fict_obj)

}
\concept{practice-crop}
