% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herd_rearing_param_sheep.R
\name{f_herd_rearing_param_sheep}
\alias{f_herd_rearing_param_sheep}
\title{Function to estimate rearing parameters of the sheep herd based on FADN data
\code{f_herd_rearing_param_sheep} Estimate rearing parameters of the sheep herd}
\usage{
f_herd_rearing_param_sheep(object)
}
\arguments{
\item{object}{a FADN2Footprint object}
}
\value{
A tibble with, for each farm, the observed quantities, flows, and rearing
parameters for the two sheep livestock categories:
\itemize{
\item LEWEBRE: breeding ewes
\item LSHEPOTH: other sheep (juveniles / fattening)
}
}
\description{
Function to estimate rearing parameters of the sheep herd based on FADN data
\code{f_herd_rearing_param_sheep} Estimate rearing parameters of the sheep herd
}
\examples{
data(fadn_fict)
fadn_fict_obj = data_4FADN2Footprint(fadn_fict)
f_herd_rearing_param_sheep(object = fadn_fict_obj)

}
\concept{practice-herding}
