% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FADN2Footprint_constructor.R
\name{f_infer_fadn_version}
\alias{f_infer_fadn_version}
\title{Infer input data base variable name version}
\usage{
f_infer_fadn_version(df, var_dict)
}
\arguments{
\item{df}{A FADN input data frame.}

\item{var_dict}{The FADN variable dictionary.}
}
\value{
The input data version.
}
\description{
Infer input data base variable name version
}
\keyword{internal}
