% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoherd_cattle.R
\name{f_pseudoherd_cattle}
\alias{f_pseudoherd_cattle}
\title{Function to estimate off-farm animals of the cattle herd based on FADN data
\code{f_pseudoherd_cattle} Estimate off-farm animals of the cattle herd}
\usage{
f_pseudoherd_cattle(object, overwrite = FALSE)
}
\arguments{
\item{object}{An S4 object of class "FADN2Footprint" prepared by the package
workflow. The function expects the object to provide:
\itemize{
\item object@traceability$id_cols (vector of id column names used for joins),
\item object@herd (for sheep, goats, horses),
\item helper functions f_herd_rearing_param_cattle, f_herd_rearing_param_swine,
f_herd_rearing_param_poultry to supply rearing parameter tables,
\item package data table data_extra$livestock for species lookup.
}}

\item{overwrite}{Logical (default FALSE). If FALSE and cached results exist,
the function returns
the cached object and no recomputation is performed. If TRUE, existing
cached GHGE results are ignored and computations are re-run.}
}
\value{
A list with, for each livestock category:
\itemize{
\item Q_obs: the number of animals observed in the farm (in number of heads)
\item Q_obs_pseudo: the number of animals of the pseudo-herd (in number of heads)
}
}
\description{
Function to estimate off-farm animals of the cattle herd based on FADN data
\code{f_pseudoherd_cattle} Estimate off-farm animals of the cattle herd
}
\examples{
data(fadn_fict)
fadn_fict_obj = data_4FADN2Footprint(fadn_fict)
f_pseudoherd_cattle(object = fadn_fict_obj)

}
\concept{practice-pseudoherd}
