% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoherd_output_econ_alloc.R
\name{f_pseudoherd_output_econ_alloc}
\alias{f_pseudoherd_output_econ_alloc}
\title{Function to estimate economic allocation ratios between herd outputs
\code{f_herd_output_econ_alloc} Estimate herd outputs}
\usage{
f_pseudoherd_output_econ_alloc(object)
}
\arguments{
\item{object}{a FADN2Footprint object}
}
\value{
A list with, for each livestock category:
\itemize{
\item Q_obs: the number of animals observed in the farm (in number of heads)
\item Q_obs_pseudo: the number of animals of the pseudo-herd (in number of heads)
}
}
\description{
Function to estimate economic allocation ratios between herd outputs
\code{f_herd_output_econ_alloc} Estimate herd outputs
}
\examples{
data(fadn_fict)
fadn_fict_obj = data_4FADN2Footprint(fadn_fict)
f_herd_output_econ_alloc(object = fadn_fict_obj)

}
\concept{practice-pseudoherd}
