% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoherd_sheep.R
\name{f_pseudoherd_sheep}
\alias{f_pseudoherd_sheep}
\title{Function to estimate off-farm animals of the sheep herd based on FADN data
\code{f_pseudoherd_sheep} Estimate off-farm animals of the sheep herd}
\usage{
f_pseudoherd_sheep(object, overwrite = FALSE)
}
\arguments{
\item{object}{a FADN2Footprint object}

\item{overwrite}{logical, force recomputation}
}
\value{
A list with rearing parameters and the pseudo herd (in number of heads)
}
\description{
Function to estimate off-farm animals of the sheep herd based on FADN data
\code{f_pseudoherd_sheep} Estimate off-farm animals of the sheep herd
}
\examples{
data(fadn_fict)
fadn_fict_obj = data_4FADN2Footprint(fadn_fict)
f_pseudoherd_sheep(object = fadn_fict_obj)

}
\concept{practice-pseudoherd}
