% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GWP}
\alias{GWP}
\title{Reference Global Warming Potentials used for conversion to CO2 equivalents}
\format{
A named vector.
}
\source{
Built from Table 8.A.1 in https://www.ipcc.ch/site/assets/uploads/2018/02/WG1AR5_Chapter08_FINAL.pdf
}
\usage{
GWP
}
\description{
\code{GWP} is an \strong{internal} auxiliary vector distributed with the
\strong{FADN2Footprint} package.
}
\details{
This dataset is intended for internal package use.
}
\examples{
# Inspect vector
data(GWP)
GWP

}
\references{
Myhre, G., et al. (2013). \emph{Anthropogenic and natural radiative forcing}, in:
\emph{Climate Change 2013: The Physical Science Basis. Contribution of Working Group I to the Fifth Assessment Report of the Intergovernmental Panel on Climate Change}. Cambridge University Press, Cambridge, United Kingdom and New York, NY, USA.
}
\keyword{datasets}
\keyword{internal}
