% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FADN2Footprint_methods.R
\name{m_remove_farms}
\alias{m_remove_farms}
\title{Remove farms from a FADN2Footprint object}
\usage{
m_remove_farms(object, farms_to_remove)
}
\arguments{
\item{object}{A \code{FADN2Footprint} object.}

\item{farms_to_remove}{A \code{data.frame} or \code{tibble} containing at least
the ID columns defined in \code{object@traceability$id_cols}. It may optionally
contain a \code{problem} column describing why the farm is being removed.}
}
\value{
The pruned \code{FADN2Footprint} object.
}
\description{
This function filters out specific farms from all data slots (farm, crop, herd,
inputs, practices, footprints, etc.) of a FADN2Footprint object. It automatically
handles nested lists and updates the traceability slot to log discarded farms.
}
\examples{
\dontrun{
# Define farms to remove (e.g., outliers)
outliers <- data.frame(ID = c(101, 105), YEAR = c(2020, 2020), problem = "Outlier in GHGE")

# Remove them
obj_clean <- m_remove_farms(my_obj, outliers)
}
}
\concept{utils}
