% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rica_fict}
\alias{rica_fict}
\title{RICA database}
\format{
A list of four dataframes: rica, veg, ani, pan
}
\source{
Ministère De L’Agriculture - SSP, 2021. Réseau d’Information Comptable Agricole - 2020. https://doi.org/10.34724/CASD.66.4159.V1
}
\usage{
data(rica_fict)
}
\description{
A Fictitious RICA database (each column is generated based on the mean and the standard deviation of the 2020 original RICA database)
}
\examples{
data(rica_fict)
str(rica_fict)
}
\keyword{datasets}
\keyword{internal}
