% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BVIAS_data.R
\name{data_for_BVIAS}
\alias{data_for_BVIAS}
\title{Intermediate functions n°1 for the estimation of BVI for crops
\code{data_for_BVIAS} Create a tibble with the following variables: "farm_id", "crop",  "land_use_type", "metric_code", "value", "max", "x_norm", "org_farming"}
\usage{
data_for_BVIAS(object, ...)
}
\arguments{
\item{object}{a FADN2Footprint object}
}
\value{
a tibble
}
\description{
Intermediate functions n°1 for the estimation of BVI for crops
\code{data_for_BVIAS} Create a tibble with the following variables: "farm_id", "crop",  "land_use_type", "metric_code", "value", "max", "x_norm", "org_farming"
}
\examples{
data(fadn_fict)
fadn_fict_obj = data_4FADN2Footprint(fadn_fict)
data_for_BVIAS(object = fadn_fict_obj)
}
\concept{footprint-biodiv}
