% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EF_electricity}
\alias{EF_electricity}
\title{GHG emission factors for electricity consumption}
\format{
a data frame (tibble) avec X lignes et 3 variables :
\describe{
  \item{COUNTRY}{Country code}
  \item{YEAR}{Years from 1990 to 2021}
  \item{EF_elec}{CoM emission factors for national electricity for EU member states, Iceland and Norway: Activity-based (IPCC) approach, CO2 emissions in tonnes CO2/MWh}
}
}
\source{
Bastos, Joana; Monforti-Ferrario, Fabio; Melica, Giulia (2024):
GHG Emission Factors for Electricity Consumption. European Commission,
Joint Research Centre (JRC) [Dataset] PID: \url{http://data.europa.eu/89h/919df040-0252-4e4e-ad82-c054896e1641}
}
\usage{
data(electricity_factors)
}
\description{
A dataset containing greenhouse gas emission factors
for electricity by country and year, produced by the JRC.
}
\keyword{datasets}
