% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed_value.R
\name{f_feed_value}
\alias{f_feed_value}
\title{Function to estimate feed value
`f_feed_value` Estimate the feed value based on EUROSTAT feed stuff prices}
\usage{
f_feed_value(object)
}
\arguments{
\item{my_object}{a FADN2Footprint object}
}
\value{
a list of two tables
}
\description{
Function to estimate feed value
`f_feed_value` Estimate the feed value based on EUROSTAT feed stuff prices
}
\examples{
data(fadn_fict)
fadn_fict_obj = data_4FADN2Footprint(fadn_fict)
f_feed_value(object = fadn_fict_obj)

}
\concept{practice-herding}
\keyword{internal}
