% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herd_rearing_param_poultry.R
\name{f_herd_rearing_param_poultry}
\alias{f_herd_rearing_param_poultry}
\title{Function to estimate off-farm animals of the poultry herd based on FADN data
`f_herd_rearing_param_poultry` Estimate off-farm animals of the poultry herd}
\usage{
f_herd_rearing_param_poultry(object)
}
\arguments{
\item{object}{a FADN2Footprint object}
}
\value{
A tibble with, for each livestock category:
- Q_obs: the number of animals observed in the farm (in number of heads)
- Q_obs_pseudo: the number of animals of the pseudo-herd (in number of heads)
}
\description{
Function to estimate off-farm animals of the poultry herd based on FADN data
`f_herd_rearing_param_poultry` Estimate off-farm animals of the poultry herd
}
\examples{
data(fadn_fict)
fadn_fict_obj = data_4FADN2Footprint(fadn_fict)
f_herd_rearing_param_poultry(object = fadn_fict_obj)

}
\concept{practice-herding}
