% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice_livestock_density.R
\name{f_livestock_density}
\alias{f_livestock_density}
\title{Function to estimate the farm livestock density based on FADN data
`livestock_density` Create a tibble with the following variables:  'farm_id', 'crop', 'TFI_ha'}
\usage{
f_livestock_density(object, overwrite = FALSE)
}
\arguments{
\item{object}{a FADN2Footprint object}

\item{overwrite}{Logical (default FALSE). If FALSE and cached results exist,
the function returns
the cached object and no recomputation is performed. If TRUE, existing
cached GHGE results are ignored and computations are re-run.}
}
\value{
a tibble
}
\description{
Function to estimate the farm livestock density based on FADN data
`livestock_density` Create a tibble with the following variables:  'farm_id', 'crop', 'TFI_ha'
}
\examples{
data(fadn_fict)
fadn_fict_obj = data_4FADN2Footprint(fadn_fict)
livestock_density(object = fadn_fict_obj)

}
\concept{practice-herding}
