% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FADN2Footprint_constructor.R
\name{f_map_columns_to_common}
\alias{f_map_columns_to_common}
\title{Map (and aggregate if needed) input data frame columns to common format}
\usage{
f_map_columns_to_common(
  df,
  id_cols = c("ID", "YEAR", "COUNTRY"),
  var_dict = dict_FADN,
  version = "var_user",
  trace = TRUE
)
}
\arguments{
\item{df}{A data.frame or tibble containing raw data (e.g., RICA or FADN).}

\item{id_cols}{A vector with the column name of the data set identifiers. Default are ID and YEAR.}

\item{var_dict}{A tibble with columns:
- `var_common`: standardized variable names (targets).
- `var_user`: original variable names (sources).}

\item{version}{One of "var_pre2014", "var_post2014", "var_user", or "var_common".}

\item{trace}{Logical; if TRUE, stores mapping info in attributes:
- `attr(df_out, "dropped_common")`: dropped common variables when source variables have matched with more than one common variable, and were stored under the first common variable name in dictionary order.
- `attr(df_out, "trace_map")`: mapping info.}
}
\value{
A tibble with harmonized variables aggregated as needed.
}
\description{
This function harmonizes a dataset (FADN or RICA) to a standardized common
variable naming scheme. It aggregates multiple source variables into one
when required (e.g., RICA's "Spring barley" + "Winter barley" -> FADN "Barley").
}
\details{
When multiple common targets exist for the same source variable
(e.g., RICA's "eggs" -> FADN "eggs for consumption" + "eggs for hatching"),
the first one (in dictionary order) is kept and the others are set to zero or empty.
}
\keyword{internal}
