% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice_mean_field_size.R
\name{f_mean_field_size}
\alias{f_mean_field_size}
\title{Function to estimate hedge density based on FADN and LPIS data
`mean_field_size` Create a tibble with the following variables:  'farm_id', 'land_use_type', 'mean_field_size'}
\usage{
f_mean_field_size(object, overwrite = FALSE)
}
\arguments{
\item{object}{a FADN2Footprint object}

\item{overwrite}{Logical (default FALSE). If FALSE and cached results exist,
the function returns
the cached object and no recomputation is performed. If TRUE, existing
cached GHGE results are ignored and computations are re-run.}
}
\value{
A tibble with:
- mean_field_size: average of hedge linear meter per hectare
}
\description{
Function to estimate hedge density based on FADN and LPIS data
`mean_field_size` Create a tibble with the following variables:  'farm_id', 'land_use_type', 'mean_field_size'
}
\examples{
data(fadn_fict)
fadn_fict_obj = data_4FADN2Footprint(fadn_fict)
mean_field_size(object = fadn_fict_obj)

}
\concept{practice-crop}
