% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoherd.R
\name{f_pseudoherd}
\alias{f_pseudoherd}
\title{Function to estimate off-farm animals of the cattle herd based on FADN data
`f_pseudoherd` Estimate off-farm animals of the cattle herd}
\usage{
f_pseudoherd(object)
}
\arguments{
\item{object}{a FADN2Footprint object}
}
\value{
A list with, for each livestock category:
- Q_obs: the number of animals observed in the farm (in number of heads)
- Q_obs_pseudo: the number of animals of the pseudo-herd (in number of heads)
}
\description{
Function to estimate off-farm animals of the cattle herd based on FADN data
`f_pseudoherd` Estimate off-farm animals of the cattle herd
}
\examples{
data(fadn_fict)
fadn_fict_obj = data_4FADN2Footprint(fadn_fict)
f_pseudoherd(object = fadn_fict_obj)

}
\concept{practice-pseudoherd}
