% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoherd_swine.R
\name{f_pseudoherd_swine}
\alias{f_pseudoherd_swine}
\title{Function to estimate off-farm animals of the swine herd based on FADN data
\code{f_pseudoherd_swine} Estimate off-farm animals of the swine herd}
\usage{
f_pseudoherd_swine(object)
}
\arguments{
\item{object}{a FADN2Footprint object}
}
\value{
A tibble with, for each livestock category:
\itemize{
\item Q_obs: the number of animals observed in the farm (in number of heads)
\item Q_obs_pseudo: the number of animals of the pseudo-herd (in number of heads)
}
}
\description{
Function to estimate off-farm animals of the swine herd based on FADN data
\code{f_pseudoherd_swine} Estimate off-farm animals of the swine herd
}
\examples{
data(fadn_fict)
fadn_fict_obj = data_4FADN2Footprint(fadn_fict)
f_pseudoherd_swine(object = fadn_fict_obj)

}
\concept{practice-pseudoherd}
