% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice_share_ferti.R
\name{f_share_Nmin_ferti}
\alias{f_share_Nmin_ferti}
\title{Function to estimate the share of fertilizer based on either FADN or RICA data
`share_Nmin_ferti` Create a tibble with 4 variables: 'farm_id', 'land_use_type', 'crop', 'A.4.3'}
\usage{
f_share_Nmin_ferti(object, overwrite = FALSE)
}
\arguments{
\item{object}{a FADN2Footprint object}

\item{overwrite}{Logical (default FALSE). If FALSE and cached results exist,
the function returns
the cached object and no recomputation is performed. If TRUE, existing
cached GHGE results are ignored and computations are re-run.}
}
\value{
a tibble
}
\description{
Function to estimate the share of fertilizer based on either FADN or RICA data
`share_Nmin_ferti` Create a tibble with 4 variables: 'farm_id', 'land_use_type', 'crop', 'A.4.3'
}
\examples{
data(fadn_fict)
fadn_fict_obj = data_4FADN2Footprint(fadn_fict)
share_Nmin_ferti(object = fadn_fict_obj)

}
\concept{practice-crop}
