% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FADN2Footprint-package.R
\docType{package}
\name{FADN2Footprint-package}
\alias{FADN2Footprint}
\alias{FADN2Footprint-package}
\title{FADN2Footprint: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
