% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FADN2Footprint_helpers.R
\name{h_read_dictionary}
\alias{h_read_dictionary}
\title{Read a filled conversion dictionary from Excel}
\usage{
h_read_dictionary(dictionary_path)
}
\arguments{
\item{dictionary_path}{Path to the filled Excel workbook.}
}
\value{
A list with elements:
\describe{
  \item{variable_pattern}{Tibble of suffix mappings by group}
  \item{category_pattern}{Tibble of product code mappings}
  \item{manual_matches}{Tibble of farm-level variable mappings}
  \item{unit_conversion}{Tibble of code-specific unit conversions}
}
}
\description{
Read a filled conversion dictionary from Excel
}
\concept{utils}
