% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_practices.R
\name{infer_practices}
\alias{infer_practices}
\title{Function to estimate agricultural practices based on FADN data
\code{infer_practices} Create a tibble with 4 variables: 'farm_id', 'land_use_type', 'crop', 'A.4.3'}
\usage{
infer_practices(object, overwrite = FALSE)
}
\arguments{
\item{object}{An object of class \code{\link{FADN2Footprint}}.}

\item{overwrite}{Logical (default FALSE). If FALSE and cached results exist,
the function returns
the cached object and no recomputation is performed. If TRUE, existing
cached GHGE results are ignored and computations are re-run.}
}
\value{
a tibble
}
\description{
Function to estimate agricultural practices based on FADN data
\code{infer_practices} Create a tibble with 4 variables: 'farm_id', 'land_use_type', 'crop', 'A.4.3'
}
\examples{
data(fadn_fict)
fadn_fict_obj = data_4FADN2Footprint(fadn_fict)
infer_practices(object = fadn_fict_obj)

}
\concept{practice-crop}
\concept{practice-herding}
